+++ src/actions.js
@@ -1,0 +1,8 @@
+/*
+ * Project: redux-i18n
+ * File: actions.js
+ */
+
+export function setLanguage(lang) {
+  return {type: "REDUX_I18N_SET_LANGUAGE", lang}
+}